packages/instructions/pyproject.toml
[project]
name = "instructions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "markdown>=3.8.2",
    "psycopg2-binary>=2.9.10",
    "ext-model",
    "markitdown[pdf]>=0.1.3",
]

[tool.uv.sources]
ext-model = { workspace = true }
